Refactor redirect checking into redirects.go .
